requirements.txt
python-dotenv
langchain 
langchain-community 
langchain-openai
langchain-ollama
#pypdf
#faiss-cpu
chromadb==0.4.22
groq
langchain-groq
beautifulsoup4
streamlit
# langchain_google_genai
PyPDF2
langchain-chroma
#pdfplumber
memory_profiler
PyMuPDF
stqdm
sentence-transformers==2.2.0
huggingface-hub
#pysqlite3-binary 
gspread
oauth2client
pandas
google-api-python-client
pytesseract
#google-generativeai
pillow
numpy
streamlit
opencv-python-headless
textblob
scipy
nltk
scikit-learn
pydantic
werkzeug
python-docx
openpyxl
fastembed==0.1.1